Pull wire-related interfaces into `wire-common/interface.ts`.
