+++ packages/data-model/src/json-wire/FabricInstanceHandler.ts
@@ -2,8 +2,8 @@ import {
   DECONSTRUCT,
   FabricInstance,
   type FabricValue,
-  type ReconstructionContext,
 } from "../interface.ts";
+import type { ReconstructionContext } from "../wire-common/interface.ts";
 import { ExplicitTagValue } from "../fabric-instances/ExplicitTagValue.ts";
 import type {
   JsonWireValue,
