+++ packages/data-model/src/json-wire/RegExpHandler.ts
@@ -1,4 +1,5 @@
-import type { FabricValue, ReconstructionContext } from "../interface.ts";
+import type { FabricValue } from "../interface.ts";
+import type { ReconstructionContext } from "../wire-common/interface.ts";
 import { FabricRegExp } from "../fabric-primitives/FabricRegExp.ts";
 import { WIRE_TYPE_TAGS } from "../wire-common/wire-type-tags.ts";
 import type {
