+++ packages/data-model/src/json-wire/json-encoding.ts
@@ -1,6 +1,6 @@
 import { isInstance } from "@commonfabric/utils/types";
 import type { FabricValue } from "../fabric-value.ts";
-import type { ReconstructionContext } from "../fabric-value.ts";
+import type { ReconstructionContext } from "../wire-common/interface.ts";
 import { EmptyReconstructionContext } from "../EmptyReconstructionContext.ts";
 import { JsonEncodingContext } from "./JsonEncodingContext.ts";
